Move InstrumentationContext api to bootstrap
